requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Image processing
Pillow>=10.0.0
imagehash>=4.3.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Web UI
streamlit>=1.28.0

# HTML reporting
jinja2>=3.1.0

# AI Summarization
google-generativeai>=0.3.0
reportlab>=4.0.0
python-dotenv>=1.0.0

# Semantic comparison (optional - for pdf_compare.pro module only)
# Can be skipped if only using Basic or Side-by-Side modes
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0

# Development and testing
pytest>=7.4.0
